addEffort {SOFIA}	R Documentation

Add Effort

Description

Add effort column to catch data.

Usage


addEffort(catch, effort, same.effort)

Arguments

catch
tibble containing year, stock, and capture.

effort
tibble containing year, stock, and effort.

same.effort
whether to use the same effort data for all stocks.

Details

If same.effort = TRUE then the effort data for stock$All is used for all stocks.

If same.effort = FALSE then the catch and effort data are paired by year and stock.

Value

Data frame like catch but with the additional column effort.

Author(s)

Arni Magnusson.

See Also

merge is the underlying function used to merge the two data frames.

addIndex adds an index column to catch data.

addDriors adds a driors column to stocks.

SOFIA-package gives an overview of the package.

Examples


## Not run: 
addEffort(catch, effort, same.effort=TRUE)
addEffort(catch, effort, same.effort=FALSE)

## End(Not run)


[Package SOFIA version 2.1.1 Index]