addIndex {SOFIA}	R Documentation

Add Index

Description

Add index column to catch data.

Usage


addIndex(catch, index, same.index)

Arguments

catch
tibble containing year, stock, and capture.

index
tibble containing year, stock, and index.

same.index
whether to use the same index data for all stocks.

Details

If same.index = TRUE then the index data for stock$All is used for all stocks.

If same.index = FALSE then the catch and index data are paired by year and stock.

Value

Data frame like catch but with the additional column index.

Author(s)

Arni Magnusson.

See Also

merge is the underlying function used to merge the two data frames.

addEffort adds an effort column to catch data.

addDriors adds a driors column to stocks.

SOFIA-package gives an overview of the package.

Examples


## Not run: 
addIndex(catch, index, same.index=TRUE)
addIndex(catch, index, same.index=FALSE)

## End(Not run)


[Package SOFIA version 2.1.1 Index]