gitClone {SOFIA}	R Documentation

Clone

Description

Clone a repository from GitHub.

Usage


gitClone(repo, topdir = "c:/git/sofia-taf", method = "https",
  tree = TRUE)

Arguments

repo
GitHub repository of the form "repo" or "owner/repo".

topdir
local directory above the repository.

method
cloning method, either "https" or "ssh".

tree
whether to organize repositories in a nested SOFIA tree, see details below.

Details

The implicit default repo owner is sofia-taf. In other words, if the user specifies a "repo" instead of "owner/repo", this function will automatically convert it to "sofia-taf/repo".

Generally, the default method = "https" works best in Windows, while method = "ssh" works best in Linux.

When tree = TRUE, this function will recognize standard SOFIA repository names such as 2022Area37Demo and clone to a nested subdirectory 2022/Area37/Demo instead of 2022Area37Demo. For repository names that do not follow this YearAreaAnalysis pattern, the value of tree has no effect.

Value

String containing the cloning command.

Author(s)

Arni Magnusson.

See Also

gitRepos fetches a complete list of GitHub repositories.

gitCloneAll clones all SOFIA-TAF repositories from GitHub.

SOFIA-package gives an overview of the package.

Examples


## Not run: 
gitClone("2022Area37Demo")
gitClone("2022Area37Demo", "~/git/sofia-taf", method="ssh")
gitClone("arni-magnusson/areaplot", "~/git/arni-magnusson", method="ssh")

## End(Not run)


[Package SOFIA version 2.1.1 Index]